+++ RxGRDB/DatabaseRegionConvertible+Rx.swift
@@ -186,7 +186,8 @@ extension Reactive where Base: DatabaseReader {
         }
         return ValueObservation
             .tracking(region, fetch: values)
-            .rx.start(
+            .rx
+            .start(
                 in: base,
                 startImmediately: startImmediately,
                 scheduler: scheduler)
@@ -202,4 +203,95 @@ extension Array where Element == DatabaseRegion {
             return DatabaseRegion()
         }
     }
+}
+
+private class ChangesObservable : ObservableType {
+    typealias E = Database
+    let writer: DatabaseWriter
+    let startImmediately: Bool
+    let observedRegion: (Database) throws -> DatabaseRegion
+    
+    /// Creates an observable that emits writer database connections on their
+    /// dedicated dispatch queue when a transaction has modified the database
+    /// in a way that impacts some requests' selections.
+    ///
+    /// When the `startImmediately` argument is true, the observable also emits
+    /// a database connection, synchronously.
+    init(
+        writer: DatabaseWriter,
+        startImmediately: Bool,
+        observedRegion: @escaping (Database) throws -> DatabaseRegion)
+    {
+        self.writer = writer
+        self.startImmediately = startImmediately
+        self.observedRegion = observedRegion
+    }
+    
+    func subscribe<O>(_ observer: O) -> Disposable where O : ObserverType, O.E == Database {
+        do {
+            let writer = self.writer
+            
+            let transactionObserver = try writer.unsafeReentrantWrite { db -> DatabaseRegionObserver in
+                if startImmediately {
+                    observer.onNext(db)
+                }
+                
+                let transactionObserver = try DatabaseRegionObserver(
+                    observedRegion: observedRegion(db),
+                    onChange: { observer.onNext(db) })
+                db.add(transactionObserver: transactionObserver)
+                return transactionObserver
+            }
+            
+            return Disposables.create {
+                writer.unsafeReentrantWrite { db in
+                    db.remove(transactionObserver: transactionObserver)
+                }
+            }
+        } catch {
+            observer.onError(error)
+            return Disposables.create()
+        }
+    }
+}
+
+private class DatabaseRegionObserver : TransactionObserver {
+    var changed: Bool = false
+    let observedRegion: DatabaseRegion
+    let change: () -> Void
+    
+    init(observedRegion: DatabaseRegion, onChange change: @escaping () -> Void) {
+        self.observedRegion = observedRegion
+        self.change = change
+    }
+    
+    func observes(eventsOfKind eventKind: DatabaseEventKind) -> Bool {
+        return observedRegion.isModified(byEventsOfKind:eventKind)
+    }
+    
+    func databaseDidChange(with event: DatabaseEvent) {
+        if observedRegion.isModified(by: event) {
+            changed = true
+            stopObservingDatabaseChangesUntilNextTransaction()
+        }
+    }
+    
+    func databaseWillCommit() { }
+    
+    func databaseDidCommit(_ db: Database) {
+        // Avoid reentrancy bugs
+        let changed = self.changed
+        self.changed = false
+        if changed {
+            change()
+        }
+    }
+    
+    func databaseDidRollback(_ db: Database) {
+        changed = false
+    }
+    
+    #if SQLITE_ENABLE_PREUPDATE_HOOK
+    func databaseWillChange(with event: DatabasePreUpdateEvent) { }
+    #endif
 }
